chore: testing to see if this fixes router circ dependencies
